// src/contracts/generated/maverick-schema.d.ts
export interface components {
  schemas: {
// ...
    Cart: {
      uuid: string;
      emailAddress: string;
      phoneNumber?: string;
      /** Format: int64 */
      eventId: number;
      /** Format: int64 */
      listingId: number;
      /** Format: int32 */
      quantity: number;
      firstName?: string;
      lastName?: string;
      pricePerTicket: number;
      serviceFee: number;
      deliveryFee: number;
      taxAmount: number;
      availableDeliveryMethods?: components['schemas']['DeliveryMethod'][];
      deliveryMethod: components['schemas']['DeliveryMethod'];
      promoDiscount: number;
      promoDescription?: string;
      giftCardRedemptions?: components['schemas']['GiftCardRedemption'][];
      section: string;
      row: string;
      notes: string;
      seatNumbers?: string;
      event: components['schemas']['Event'];
      emailOptIn: boolean;
      smsOptIn: boolean;
      listing: components['schemas']['Listing'];
      shipToCountryCode?: string;
      /** Format: int64 */
      orderId?: number;
      checkoutVersion?: string;
      experimentId?: string;
      /** Format: int32 */
      variationId?: number;
      amountToCharge?: number;
      serviceFeePerTicket?: number;
      cartTotal?: number;
    };
    Category: {
      /** Format: int64 */
      id: number;
      name: string;
      /** Format: int64 */
      parentId?: number;
      /** @enum {string} */
      eventType: 'CONCERTS' | 'COMEDY' | 'SPORTS' | 'THEATER' | 'PARKING';
      slug?: string;
      heroImagePath?: string;
      heroImageCredit?: string;
      cardImagePath?: string;
      performerHeroImagePath?: string;
      performerHeroImageCredit?: string;
      performerCardImagePath?: string;
    };
    DeliveryMethod: {
      /** Format: int64 */
      id: number;
      enumName: string;
      displayName: string;
      retailDisplayName: string;
      description: string;
    };
    Event: {
      /** Format: int64 */
      doc?: number;
      /** Format: int64 */
      fields?: number;
      /** Format: int64 */
      id: number;
      name: string;
      /** Format: int64 */
      venueId: number;
      /** Format: int64 */
      venueConfigurationId: number;
      /** Format: date-time */
      eventTimeLocal: string;
      /** Format: date-time */
      endDate?: string;
      timeTbd: boolean;
      dateTbd: boolean;
      url?: string;
      parking: boolean;
      /** Format: int64 */
      parkingEventId?: number;
      /** Format: int32 */
      minimumAge?: number;
      ifNecessary: boolean;
      multiDate: boolean;
      /** Format: date-time */
      eventTimeUtc?: string;
      /** Format: int64 */
      cmsPageId?: number;
      eventPerformers?: components['schemas']['EventPerformer'][];
      venue?: components['schemas']['Venue'];
      slug?: string;
      venueConfiguration?: components['schemas']['VenueConfiguration'];
      seasonTicket: boolean;
      /** Format: date-time */
      nextOnsaleTimeLocal?: string;
      /** Format: date-time */
      nextPresaleTimeLocal?: string;
      /** Format: date-time */
      lastOnsaleTimeLocal?: string;
      /** Format: date-time */
      lastPresaleTimeLocal?: string;
      /** Format: date-time */
      firstOnSaleTimeLocal?: string;
      /** Format: int32 */
      daysToEvent?: number;
      /** Format: int32 */
      daysToOnsale?: number;
      /** Format: int32 */
      daysToPresale?: number;
      /** Format: int32 */
      daysUntilFirstOnSale?: number;
      /** @enum {string} */
      pricingType?: 'DEFAULT' | 'ALL_IN_TICKET_DETAILS' | 'ALL_IN_TICKET_PAGE';
      /** Format: int32 */
      availableTickets?: number;
      /** Format: double */
      percentInventoryAvailable?: number;
    };
    EventPerformer: {
      /** Format: int64 */
      id: number;
      /** Format: int64 */
      performerId: number;
      master: boolean;
      performer?: components['schemas']['Performer'];
    };
    GiftCardRedemption: {
      amountRedeemed: number;
      /** Format: int64 */
      giftCardId: number;
    };
    Listing: {
      /** Format: int64 */
      id: number;
      /** Format: int64 */
      eventId: number;
      section: string;
      /** Format: int64 */
      sectionId: number;
      row: string;
      notes?: string;
      /** Format: int32 */
      quantity: number;
      validSplitQuantities: number[];
      flex: boolean;
      instant: boolean;
      displayPrice: number;
      serviceFee?: number;
      powerSeller: boolean;
      /** Format: date */
      inHandDate?: string;
      stockType: string;
      deliveryMethods: components['schemas']['DeliveryMethod'][];
      attributes?: number[];
      seatNumbers?: string;
      generalAdmission: boolean;
    };
    Performer: {
      /** Format: int64 */
      id: number;
      name: string;
      /** Format: int64 */
      homeVenueId?: number;
      /** Format: int64 */
      homeVenueConfigId?: number;
      /** Format: int64 */
      primaryCategoryId: number;
      /** Format: int64 */
      parentPerformerId?: number;
      parking: boolean;
      /** Format: int64 */
      parkingPerformerId?: number;
      slug?: string;
      heroImagePath?: string;
      heroImageCredit?: string;
      cardImagePath?: string;
      primaryCategory?: components['schemas']['Category'];
      /** Format: int64 */
      cmsPageId?: number;
      /** Format: int32 */
      eventCount?: number;
      parkingPerformer?: components['schemas']['Performer'];
      parentPerformer?: components['schemas']['Performer'];
      hasChildren?: boolean;
      childPerformers?: components['schemas']['Performer'][];
      homeVenue?: components['schemas']['Venue'];
      parkingParent?: components['schemas']['Performer'];
      /** Format: int32 */
      popularity?: number;
      performerStats?: components['schemas']['PerformerStats'];
    };
// ...
    Order: {
      /** Format: int64 */
      id: number;
      /** Format: date-time */
      createdAt: string;
      status: components['schemas']['OrderStatus'];
      emailAddress: string;
      phoneNumber: string;
      deliveryMethod?: string;
      deliveryMethodDetails?: components['schemas']['DeliveryMethod'];
      serviceFee: number;
      taxAmount: number;
      deliveryFee: number;
      insuranceFee: number;
      hasInsurance: boolean;
      /** @enum {string} */
      insuranceOfferStatus?: 'NO_OFFER' | 'ACCEPTED' | 'DECLINED';
      otherFee?: number;
      orderTotal: number;
      /** Format: date */
      inHandDate?: string;
      orderNotes?: string;
      delivered: boolean;
      /** Format: date-time */
      deliveryTime?: string;
      processed: boolean;
      cancelled: boolean;
      billingAddress?: components['schemas']['Address'];
      shippingAddress?: components['schemas']['Address'];
      items: components['schemas']['OrderItem'][];
      paymentMethod?: string;
      creditCardLastFour?: string;
      creditCardType?: string;
      identityVerificationUrl?: string;
      gp?: number;
      showReviewPopup?: boolean;
      insurance?: components['schemas']['OrderInsurance'];
      declinedInsurance: boolean;
      rca?: boolean;
      promoCodeLastFour?: string;
    };
// ...
    CmsPathResponse: {
      /** Format: int64 */
      pageId?: number;
      path?: string;
      title?: string;
      metaDescription?: string;
      metaKeywords?: string;
      event?: components['schemas']['Event'];
      venue?: components['schemas']['Venue'];
      region?: components['schemas']['Region'];
      performer?: components['schemas']['Performer'];
      template?: string;
      events?: components['schemas']['Event'][];
      paginatedEvents?: components['schemas']['PaginatedResultEvent'];
      customFieldValues?: {
        [key: string]: string;
      };
      topPerformers?: components['schemas']['Performer'][];
      category?: components['schemas']['Category'];
      redirect?: boolean;
      redirectLocation?: string;
      eventSearchParameters?: components['schemas']['EventSearchParameters'];
      showScarcity?: boolean;
      cookies?: components['schemas']['Cookie'][];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}